Don't modify PATH on OS X
There is a node executable in the apm/bin directory, which overrides any version of Node.js installed on the system since the PATH is being updated to put this path at the front. As the Linux side of the script does not do this, remove the modification of the PATH to stop breaking Node.js.