Don't pipe script directly to sh
On some systems (CircleCI trusty images) the curl command closes the
connection early, causing the rest of the script to be directly sent to
`stdout`... and consequently missed by `sh`, causing the execution to halt
at wherever the script got cut off. This changes the script to be
downloaded to a local file and executed from there.